requirements.txt
appdirs==1.4.4
asgiref==3.2.10
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.2
django-heroku==0.3.1
djangorestframework==3.12.2
filelock==3.0.12
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
MarkupSafe==1.1.1
numpy==1.19.1
opencv-python==4.3.0.36
Pillow==8.0.1
postgres==3.0.0
psycopg2==2.8.6
psycopg2-binary==2.8.5
psycopg2-pool==1.1
pytz==2020.1
requests==2.24.0
scipy==1.5.2
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.18
sqlparse==0.4.1
threadpoolctl==2.1.0
urllib3==1.25.11
virtualenv==20.0.30
Werkzeug==1.0.1
whitenoise==5.2.0
WTForms==2.3.3
